requirements.txt
et_xmlfile==2.0.0
Jinja2==3.1.6
MarkupSafe==3.0.3
openpyxl==3.1.5
PyYAML==6.0.3
